+++ packages/nextra/src/plugin.ts
@@ -249,28 +249,3 @@ ${(err as Error).name}: ${(err as Error).message}`
 
   return { items, fileMap }
 }
-
-export class PageMapCache {
-  cache: {
-    items: PageMapItem[]
-    fileMap: FileMap
-  } | null = {
-    items: [],
-    fileMap: Object.create(null)
-  }
-
-  set(data: { items: PageMapItem[]; fileMap: FileMap }) {
-    this.cache!.items = data.items
-    this.cache!.fileMap = data.fileMap
-  }
-
-  clear() {
-    this.cache = null
-  }
-
-  get() {
-    return this.cache
-  }
-}
-
-export const pageMapCache = new PageMapCache()
